# Faster unstacking to sparse