--- a/asv_bench/asv.conf.json
+++ b/asv_bench/asv.conf.json
@@ -65,6 +65,7 @@
         "bottleneck": ["", null],
         "dask": [""],
         "distributed": [""],
+        "sparse": [""]
     },