--- a/xarray/core/variable.py
+++ b/xarray/core/variable.py
@@ -1613,6 +1613,8 @@ class Variable(AbstractArray, NdimSizeLenMixin, VariableArithmetic):
                 data=self.data.astype(dtype).ravel(),
                 fill_value=fill_value,
                 shape=new_shape,
+                has_duplicates=False,
+                sorted=index.is_monotonic_increasing,
             )
 
         else: