--- a/xarray/tests/test_dataset.py
+++ b/xarray/tests/test_dataset.py
@@ -29,7 +29,7 @@ from xarray.coding.cftimeindex import CFTimeIndex
 from xarray.core import dtypes, indexing, utils
 from xarray.core.common import duck_array_ops, full_like
 from xarray.core.indexes import Index
-from xarray.core.pycompat import integer_types
+from xarray.core.pycompat import integer_types, sparse_array_type
 from xarray.core.utils import is_scalar
 
 from . import (
@@ -3084,13 +3084,31 @@ class TestDataset:
         # test fill_value
         actual = ds.unstack("index", sparse=True)
         expected = ds.unstack("index")
+        assert isinstance(actual["var"].data, sparse_array_type)
         assert actual["var"].variable._to_dense().equals(expected["var"].variable)
         assert actual["var"].data.density < 1.0
 
         actual = ds["var"].unstack("index", sparse=True)
         expected = ds["var"].unstack("index")
+        assert isinstance(actual.data, sparse_array_type)
         assert actual.variable._to_dense().equals(expected.variable)
         assert actual.data.density < 1.0
+
+        mindex = pd.MultiIndex.from_arrays(
+            [np.arange(3), np.arange(3)], names=["a", "b"]
+        )
+        ds_eye = Dataset(
+            {"var": (("z", "foo"), np.ones((3, 4)))},
+            coords={"z": mindex, "foo": np.arange(4)},
+        )
+        actual = ds_eye.unstack(sparse=True, fill_value=0)
+        assert isinstance(actual["var"].data, sparse_array_type)
+        expected = xr.Dataset(
+            {"var": (("foo", "a", "b"), np.broadcast_to(np.eye(3, 3), (4, 3, 3)))},
+            coords={"foo": np.arange(4), "a": np.arange(3), "b": np.arange(3)},
+        )
+        actual["var"].data = actual["var"].data.todense()
+        assert_equal(expected, actual)
 
     def test_stack_unstack_fast(self):
         ds = Dataset(