<!-- Feel free to remove check-list items aren't relevant to your change -->

- [x] Tests added
- [x] Passes `pre-commit run --all-files`
- [x] User visible changes (including notable bug fixes) are documented in `whats-new.rst`

From 7s to 25 ms and 3.5GB to 850MB memory usage =) by passing the coordinate locations directly to the sparse constructor.
```
asv run -e --bench unstacking.UnstackingSparse.time_unstack_to_sparse  --cpu-affinity=3 HEAD
[  0.00%] · For xarray commit c9251e1c <sparse-unstack>:
[  0.00%] ·· Building for conda-py3.8-bottleneck-dask-distributed-netcdf4-numpy-pandas-scipy-sparse
[  0.00%] ·· Benchmarking conda-py3.8-bottleneck-dask-distributed-netcdf4-numpy-pandas-scipy-sparse
[  0.01%] ··· Running (unstacking.UnstackingSparse.time_unstack_to_sparse_2d--)..
[  0.02%] ··· unstacking.UnstackingSparse.time_unstack_to_sparse_2d    623±30μs
[  0.02%] ··· unstacking.UnstackingSparse.time_unstack_to_sparse_3d    22.8±2ms
[  0.06%] ··· unstacking.UnstackingSparse.peakmem_unstack_to_sparse_2d    793M
[  0.06%] ··· unstacking.UnstackingSparse.peakmem_unstack_to_sparse_3d    794M


[  0.04%] · For xarray commit 80905135 <main>:
[  0.04%] ·· Building for conda-py3.8-bottleneck-dask-distributed-netcdf4-numpy-pandas-scipy-sparse..
[  0.04%] ·· Benchmarking conda-py3.8-bottleneck-dask-distributed-netcdf4-numpy-pandas-scipy-sparse
[  0.05%] ··· Running (unstacking.UnstackingSparse.time_unstack_to_sparse_2d--)..
[  0.06%] ··· unstacking.UnstackingSparse.time_unstack_to_sparse_2d    596±30ms
[  0.06%] ··· unstacking.UnstackingSparse.time_unstack_to_sparse_3d    7.72±0.1s
[  0.02%] ··· unstacking.UnstackingSparse.peakmem_unstack_to_sparse_2d    867M
[  0.02%] ··· unstacking.UnstackingSparse.peakmem_unstack_to_sparse_3d    3.56G
```

cc @bonnland