@max-sixty 2021-07-05
> From 7s to 25 ms

Casual!