@dcherian 2021-12-02
@pydata/xarray I'm planning to merge on Friday. It's been sitting around for a while and is a giant improvement.